+++ src/Fa/FaStacked.js
@@ -14,6 +14,7 @@ export const propTypes = {
     '3x',
     '4x',
     '5x',
+    '',
   ]),
 };